@@ -15,14 +15,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~13.1.1",
-    "@angular/core": "~13.1.1",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
     "@angular/fire": "^7.4.1",
-    "@angular/forms": "~13.1.1",
-    "@angular/platform-browser": "~13.1.1",
-    "@angular/platform-browser-dynamic": "~13.1.1",
-    "@angular/router": "~13.1.1",
-    "@angular/service-worker": "~13.1.1",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@angular/service-worker": "^14.3.0",
     "@capacitor/android": "3.2.5",
     "@capacitor/app": "1.0.5",
     "@capacitor/core": "3.2.5",
@@ -41,16 +41,16 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.2",
+    "@angular-devkit/build-angular": "^14.2.11",
     "@angular-eslint/builder": "13.5.0",
     "@angular-eslint/eslint-plugin": "13.5.0",
     "@angular-eslint/eslint-plugin-template": "13.5.0",
     "@angular-eslint/schematics": "13.5.0",
     "@angular-eslint/template-parser": "13.5.0",
-    "@angular/cli": "~13.1.2",
-    "@angular/compiler": "~13.1.1",
-    "@angular/compiler-cli": "~13.1.1",
-    "@angular/language-service": "~13.1.1",
+    "@angular/cli": "^14.2.11",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@capacitor/cli": "3.2.5",
     "@ionic/angular-toolkit": "^4.0.0",
     "@types/jasmine": "~3.6.0",
@@ -72,7 +72,7 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.5.4"
+    "typescript": "~4.8.4"
   },
   "description": "An Ionic project"
 }
